:recycle: refactor(#168/auth): auth 도메인 관련 클래스 ninedot-domain 으로 이동
