+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/auth/exception/AuthErrorCode.java
@@ -1,8 +1,9 @@
 package org.sopt36.ninedotserver.auth.exception;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt36.ninedotserver.global.exception.ErrorCode;
+import org.sopt36.ninedotserver.exception.ErrorCode;
 import org.springframework.http.HttpStatus;
+
 
 @RequiredArgsConstructor
 public enum AuthErrorCode implements ErrorCode {
@@ -18,13 +19,9 @@ public enum AuthErrorCode implements ErrorCode {
     GOOGLE_USER_INFO_RETRIEVAL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "구글 사용자 정보 조회에 실패했습니다."),
     GOOGLE_TOKEN_RETRIEVAL_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "구글 토큰 발급에 실패했습니다.");
 
+    @lombok.Getter
     private final HttpStatus status;
     private final String message;
-
-    @Override
-    public HttpStatus getStatus() {
-        return status;
-    }
 
     @Override
     public String getMessage() {
