+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/auth/exception/AuthException.java
@@ -1,6 +1,6 @@
 package org.sopt36.ninedotserver.auth.exception;
 
-import org.sopt36.ninedotserver.global.exception.BusinessException;
+import org.sopt36.ninedotserver.exception.BusinessException;
 
 public class AuthException extends BusinessException {
