+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/auth/model/OnboardingPage.java
@@ -1,0 +1,5 @@
+package org.sopt36.ninedotserver.auth.model;
+
+public enum OnboardingPage {
+    MANDALART, CORE_GOAL, SUB_GOAL, ONBOARDING_COMPLETED
+}
