--- src/main/java/org/sopt36/ninedotserver/auth/domain/OnboardingPage.java
@@ -1,5 +1,0 @@
-package org.sopt36.ninedotserver.auth.domain;
-
-public enum OnboardingPage {
-    MANDALART, CORE_GOAL, SUB_GOAL, ONBOARDING_COMPLETED;
-}
